refactor(wishlist): move functions into lib, refactor function names
